+++ src/Sentry.Unity.Editor/Android/AndroidManifestConfiguration.cs
@@ -13,29 +13,46 @@ namespace Sentry.Unity.Editor.Android
     public class AndroidManifestConfiguration : IPostGenerateGradleAndroidProject
     {
         private readonly Func<SentryUnityOptions?> _getOptions;
+        private readonly Func<SentryCliOptions?> _getSentryCliOptions;
         private readonly IUnityLoggerInterceptor? _interceptor;
+
+        private readonly bool _isDevelopmentBuild;
+        private readonly ScriptingImplementation _scriptingImplementation;
 
         // Lower levels are called first.
         public int callbackOrder => 1;
 
         public AndroidManifestConfiguration()
-            : this(() => ScriptableSentryUnityOptions.LoadSentryUnityOptions(BuildPipeline.isBuildingPlayer))
+            : this(() => ScriptableSentryUnityOptions.LoadSentryUnityOptions(BuildPipeline.isBuildingPlayer),
+                () => SentryCliOptions.LoadCliOptions(),
+                isDevelopmentBuild: EditorUserBuildSettings.development,
+                scriptingImplementation: PlayerSettings.GetScriptingBackend(BuildTargetGroup.Android))
         {
         }
 
         // Testing
         internal AndroidManifestConfiguration(
             Func<SentryUnityOptions?> getOptions,
-            IUnityLoggerInterceptor? interceptor = null)
+            Func<SentryCliOptions?> getSentryCliOptions,
+            IUnityLoggerInterceptor? interceptor = null,
+            bool isDevelopmentBuild = false,
+            ScriptingImplementation scriptingImplementation = ScriptingImplementation.IL2CPP)
         {
             _getOptions = getOptions;
+            _getSentryCliOptions = getSentryCliOptions;
             _interceptor = interceptor;
+
+            _isDevelopmentBuild = isDevelopmentBuild;
+            _scriptingImplementation = scriptingImplementation;
         }
 
         public void OnPostGenerateGradleAndroidProject(string basePath)
         {
             ModifyManifest(basePath);
-            SetupSymbolsUpload(basePath);
+
+            var unityProjectPath = Directory.GetParent(Application.dataPath).FullName;
+            var gradleProjectPath = Directory.GetParent(basePath).FullName;
+            SetupSymbolsUpload(unityProjectPath, gradleProjectPath);
         }
 
         internal void ModifyManifest(string basePath)
@@ -128,22 +145,33 @@ namespace Sentry.Unity.Editor.Android
             _ = androidManifest.Save();
         }
 
-        internal void SetupSymbolsUpload(string basePath)
+        internal void SetupSymbolsUpload(string unityProjectPath, string gradleProjectPath)
         {
             var options = _getOptions();
             var logger = options?.DiagnosticLogger ?? new UnityLogger(new SentryUnityOptions());
 
-            var sentryCliOptions = SentryCliOptions.LoadCliOptions();
+            if (_scriptingImplementation != ScriptingImplementation.IL2CPP)
+            {
+                logger.LogDebug("Automated symbols upload requires the IL2CPP scripting backend.");
+                return;
+            }
+
+            var sentryCliOptions = _getSentryCliOptions();
+            if (sentryCliOptions is null)
+            {
+                logger.LogWarning("Failed to load sentry-cli options - Skipping symbols upload.");
+                return;
+            }
+
             if (!sentryCliOptions.UploadSymbols)
             {
                 logger.LogDebug("Automated symbols upload has been disabled.");
                 return;
             }
 
-            if (EditorUserBuildSettings.development && !sentryCliOptions.UploadDevelopmentSymbols)
+            if (_isDevelopmentBuild && !sentryCliOptions.UploadDevelopmentSymbols)
             {
-                logger.LogDebug(
-                    "Automated symbols upload for development builds has been disabled.");
+                logger.LogDebug("Automated symbols upload for development builds has been disabled.");
                 return;
             }
 
@@ -157,9 +185,6 @@ namespace Sentry.Unity.Editor.Android
 
             try
             {
-                var unityProjectPath = Directory.GetParent(Application.dataPath).FullName;
-                var gradleProjectPath = Directory.GetParent(basePath).FullName;
-
                 var symbolsPath = DebugSymbolUpload.GetSymbolsPath(
                     unityProjectPath,
                     gradleProjectPath,