+++ src/Sentry.Unity/SentryOptionsUtility.cs
@@ -6,32 +6,9 @@ namespace Sentry.Unity
 {
     internal static class SentryOptionsUtility
     {
-        public static void SetDefaults(SentryUnityOptions options, IApplication? application = null,
-            bool isBuilding = false)
-        {
-            application ??= ApplicationAdapter.Instance;
-
-            options.Enabled = true;
-            options.IsGlobalModeEnabled = true;
-
-            options.AutoSessionTracking = true;
-            options.CaptureInEditor = true;
-            options.RequestBodyCompressionLevel = CompressionLevelWithAuto.NoCompression;
-            options.InitCacheFlushTimeout = System.TimeSpan.Zero;
-
-            options.StackTraceMode = StackTraceMode.Original;
-            options.IsEnvironmentUser = false;
-
-            options.Release = Release(application);
-            options.Environment = Environment(application, isBuilding);
-
-            options.CacheDirectoryPath = application.PersistentDataPath;
-        }
-
         public static void SetDefaults(ScriptableSentryUnityOptions scriptableOptions)
         {
             var options = new SentryUnityOptions();
-            SetDefaults(options);
 
             scriptableOptions.Enabled = options.Enabled;
 
@@ -66,21 +43,6 @@ namespace Sentry.Unity
             scriptableOptions.DebugOnlyInEditor = true;
             scriptableOptions.DiagnosticLevel = SentryLevel.Warning;
         }
-
-        private static string Release(IApplication application)
-        {
-            if (application.ProductName is string productName
-                && !string.IsNullOrWhiteSpace(productName)
-                && productName.Any(c => c != '.')) // productName consisting solely of '.'
-            {
-                productName = Regex.Replace(productName, @"\n|\r|\t|\/|\\|\.{2}|@", "_");
-                return $"{productName}@{application.Version}";
-            }
-
-            return application.Version;
-        }
-
-        private static string Environment(IApplication application, bool isBuilding) => (application.IsEditor && !isBuilding) ? "editor" : "production";
 
         public static void TryAttachLogger(SentryUnityOptions options, IApplication? application = null)
         {